fix(P0): 商城5个严重Bug修复
1. shop/detail.php→view.php: 3处broken link修复(mall/index+product/detail+shop/list)
2. getUserShopStatus: shop/list.php改为内联Shop::userHasShop()
3. seed.php: 加admin角色检查
4. BCT支付: intval→floatval保精度
5. 立即购买: add_to_cart.php支持clear_first清车再买

+++ mall/user/pay.php
@@ -75,7 +75,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
                     $_SESSION['user_id'],              // from_user (买家)
                     $sellerId ?: 0,                    // to_user (卖家)
                     $orderInfo['payment_city'] ?? '',  // city
-                    intval($orderInfo['payment_amount'] ?? 0), // amount
+                    floatval($orderInfo['payment_amount'] ?? 0), // amount
                     1.0,                               // price
                     0,                                 // fee
                     null,                              // fee_type
